+++ moview/utils/data_manager.py
@@ -3,7 +3,7 @@ import sys
 from typing import *
 
 from util import remove_indent
-from core.custom_callback_handler import CustomCallbackHandler
+from moview.handlers.custom_callback_handler import CustomCallbackHandler
 from langchain.chat_models import ChatOpenAI
