+++ resources/CreateQuestionDiagram.png
